// test/fixtures/api/__routes__/myApi.d.ts
// Generated by unrouted
export interface RouteSchema {
  post:  '/post-test'|
  '/names'|
  '/blog/articles'|
  `/blog/articles/${string}`|
  '/any-route'

  get:  '/greeting'|
  `/greeting/${string}`|
  '/names'|
  '/old-link'|
  '/older-link'|
  '/new-link'|
  '/new-permalink'|
  '/blog/articles'|
  `/blog/articles/${string}`|
  '/any-route'|
  `/static/${string}`|
  '/static'

  head:  '/greeting'|
  `/greeting/${string}`|
  '/names'|
  '/new-link'|
  '/new-permalink'|
  '/blog/articles'|
  `/blog/articles/${string}`|
  '/any-route'

  put:  `/blog/articles/${string}`|
  '/any-route'

  delete:  `/blog/articles/${string}`|
  '/any-route'

  connect:  '/any-route'

  options:  '/any-route'

  trace:  '/any-route'

}